requirements.txt
alembic==1.0.6
aniso8601==4.1.0
appnope==0.1.0
Babel==2.6.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configparser==3.5.0
debtcollector==1.20.0
decorator==4.3.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.3.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
ipython==7.2.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
keystoneauth1==3.11.2
Mako==1.0.7
MarkupSafe==1.1.0
mongoengine==0.16.3
msgpack==0.6.0
netaddr==0.7.19
netifaces==0.10.9
os-service-types==1.4.0
oslo.config==6.7.0
oslo.i18n==3.23.0
oslo.serialization==2.28.1
oslo.utils==3.39.1
parso==0.3.1
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
pymongo==3.7.2
PyMySQL==0.9.3
pyparsing==2.3.0
python-dateutil==2.7.5
python-editor==1.0.3
python-keystoneclient==3.18.0
pytz==2018.9
PyYAML==3.13
redis==3.0.1
requests==2.21.0
rfc3986==1.2.0
rq==0.13.0
six==1.12.0
SQLAlchemy==1.2.15
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
stevedore==1.30.0
Tempita==0.5.2
traitlets==4.3.2
urllib3==1.24.1
validators==0.12.4
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.0
WTForms==2.2.1
